% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_clustering.R
\name{annotate_clusters}
\alias{annotate_clusters}
\title{Annotate data frame with clusters}
\usage{
annotate_clusters(df, clusters, long = TRUE, selected_clusters = NULL)
}
\arguments{
\item{df}{a data frame}

\item{clusters}{list of cluster labels, automatically converted to factor.}

\item{long}{if TRUE, returned data frame will be in long format. See details for spec.}

\item{selected_clusters}{optional labels to filter}
}
\value{
a wide or long data frame
}
\description{
Annotate data frame with clusters
}
\details{
Long data frame will have columns: \code{Cluster}, \code{Measurement} and \code{Value}.
}
\examples{
res <- compute_clusters(dist(iris[, c("Petal.Length", "Sepal.Length")]), "complete")
cluster_labels <- cut_clusters(res, 2)
annotated_data <- annotate_clusters(iris[, c("Petal.Length", "Sepal.Length")], cluster_labels)
head(annotated_data)
}
