% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmaps.R
\docType{data}
\name{cluster_colors}
\alias{cluster_colors}
\title{List of colors used in the Shiny app for clusters}
\format{
An object of class \code{character} of length 39.
}
\usage{
cluster_colors
}
\description{
List of colors used in the Shiny app for clusters
}
\keyword{datasets}
