% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmaps.R
\name{cluster_heatmaps}
\alias{cluster_heatmaps}
\title{Draw two heatmaps}
\usage{
cluster_heatmaps(
  top_matrix,
  bottom_matrix,
  dendrograms,
  clusters_set,
  annotation = NULL
)
}
\arguments{
\item{top_matrix}{matrix with selected variables}

\item{bottom_matrix}{matrix with unselected variables}

\item{dendrograms}{to draw above top matrix}

\item{clusters_set}{list of cluster indices}

\item{annotation}{any kind of annotation object to draw as top_annotation}
}
\value{
two concatenated heatmaps drawn with ComplexHeatmap::draw
}
\description{
Draw two heatmaps
}
\keyword{internal}
