% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_clustering.R
\name{compute_clusters}
\alias{compute_clusters}
\title{Compute clusters hierarchically from distance matrix}
\usage{
compute_clusters(dmat, linkage_method)
}
\arguments{
\item{dmat}{a distance matrix}

\item{linkage_method}{a linkage method supported by \code{\link[fastcluster:hclust]{fastcluster::hclust()}}}
}
\value{
clusters computed by \code{\link[fastcluster:hclust]{fastcluster::hclust()}}
}
\description{
Compute clusters hierarchically from distance matrix
}
\examples{
dmat <- compute_dmat(iris, "euclidean", TRUE, c("Petal.Length", "Sepal.Length"))
res <- compute_clusters(dmat, "complete")
}
