% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_clustering.R
\name{compute_dmat}
\alias{compute_dmat}
\title{Compute a distance matrix from scaled data}
\usage{
compute_dmat(
  x,
  dist_method = "euclidean",
  scaling_method = NULL,
  subset_cols = NULL
)
}
\arguments{
\item{x}{a numeric data frame or matrix}

\item{dist_method}{a distance measure to apply to the scaled data. Must be those supported by \code{\link[stats:dist]{stats::dist()}}, plus 'mahalanobis' and 'cosine'.}

\item{scaling_method}{method to scale data frame columns. Valid values are \code{z-scores} and \code{robust}, any other will return unscaled data.}

\item{subset_cols}{a list of columns to subset the data}
}
\value{
an object of class "dist" (see \code{\link[stats:dist]{stats::dist()}})
}
\description{
This function applies scaling to the columns of a data frame and
computes and returns a distance matrix from a chosen distance measure.
}
