% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmaps.R
\name{correlation_heatmap}
\alias{correlation_heatmap}
\title{Plot a correlation heatmap}
\usage{
correlation_heatmap(df)
}
\arguments{
\item{df}{numeric data frame to compute correlations}
}
\value{
a \link[ComplexHeatmap:Heatmap]{ComplexHeatmap::Heatmap}
}
\description{
Computes pairwise Pearson correlation; if there are fewer than 15 columns, prints
the value of the correlation coefficient inside each tile.
}
