% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmaps.R
\name{create_annotations}
\alias{create_annotations}
\title{Create heatmap annotations from selected variables}
\usage{
create_annotations(df, selected_variables)
}
\arguments{
\item{df}{a data frame. It can be an original unscaled data, or a scaled one}

\item{selected_variables}{list of columns in the data frame to create annotations for}
}
\value{
a \link[ComplexHeatmap:columnAnnotation]{ComplexHeatmap::columnAnnotation} object
}
\description{
This function will create a \link[ComplexHeatmap:columnAnnotation]{ComplexHeatmap::columnAnnotation} object with rows
for each variable passed as argument. Character columns will be coerced into factors.
For factors, the ColorBrewer palette \code{Set3} will be used. For non-negative numeric, the
\code{PuBu} palette will be used, and for columns with negative values, the reversed \code{RdBu} will be used.
}
