% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{line_plot}
\alias{line_plot}
\title{A custom line plot with optional vertical line}
\usage{
line_plot(df, x, y, xintercept = NULL)
}
\arguments{
\item{df}{data source}

\item{x}{variable for horizontal axis}

\item{y}{variable for vertical axis}

\item{xintercept}{optional value in horizontal axis to highlight}
}
\value{
a \link[ggplot2:ggplot]{ggplot2::ggplot} object
}
\description{
A custom line plot with optional vertical line
}
