% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{normal_annotated}
\alias{normal_annotated}
\title{Simulated normal data with annotations}
\format{
A data frame with 200 rows and 10 variables:
\describe{
\item{a}{variable a}
\item{b}{variable b}
\item{c}{variable c}
\item{d}{variable d}
\item{e}{variable e}
\item{f}{variable f}
\item{g}{variable g}
\item{h}{variable h}
\item{i}{variable i}
\item{j}{variable j}
\item{annot}{annotation column}
}
}
\source{
package author
}
\usage{
normal_annotated
}
\description{
Simulated normal data with annotations
}
\keyword{datasets}
