% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Runs the Shiny app}
\usage{
run_app()
}
\value{
No return value, runs the app by passing it to print
}
\description{
Runs the Shiny app
}
\examples{
## Only run this example in interactive R sessions
if (interactive()) {
library(visxhclust)
run_app()
}
}
