% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{scale_data}
\alias{scale_data}
\title{Apply a specified scaling method}
\usage{
scale_data(x, method = "z-scores")
}
\arguments{
\item{x}{a numeric data frame or matrix}

\item{method}{the scaling method. Currently supported are \code{"z-scores"} or \code{"robust"}}
}
\value{
a scaled data frame
}
\description{
Apply a specified scaling method
}
\details{
Robust scaling is done via median absolute deviation
}
\examples{
scaled_data <- scale_data(iris[, c("Petal.Length", "Sepal.Length")], method = "z-scores")
head(scaled_data)
}
