% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmaps.R
\name{plot_cluster_heatmaps}
\alias{plot_cluster_heatmaps}
\title{Draw two heatmaps}
\usage{
plot_cluster_heatmaps(
  top_matrix,
  bottom_matrix,
  dendrograms,
  clusters_set,
  annotation = NULL,
  scaled = FALSE,
  distance_method = NULL,
  cluster_features = TRUE,
  show_col_names = TRUE
)
}
\arguments{
\item{top_matrix}{matrix with selected variables}

\item{bottom_matrix}{matrix with unselected variables}

\item{dendrograms}{to draw above top matrix}

\item{clusters_set}{list of cluster indices}

\item{annotation}{(optional) any kind of annotation object to draw as top_annotation}

\item{scaled}{(optional) boolean to modify colour scale if data has already been scaled}

\item{distance_method}{(optional) if "Binary", use discrete colors for heatmap}

\item{cluster_features}{(optional) If FALSE, row order does not change}

\item{show_col_names}{(optional) If FALSE, does not show column names at base of heatmap}
}
\value{
two concatenated heatmaps drawn with ComplexHeatmap::draw
}
\description{
Draw two heatmaps
}
\keyword{internal}
